+++ src/test/java/persistence/entity/DefaultEntityPersisterTest.java
@@ -14,7 +14,7 @@ import persistence.sql.dml.DeleteQueryBuilder;
 import persistence.sql.dml.InsertQueryBuilder;
 import persistence.sql.dml.SelectQueryBuilder;
 import persistence.sql.dml.UpdateQueryBuilder;
-import persistence.sql.meta.EntityTable;
+import persistence.meta.EntityTable;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
