+++ src/test/java/persistence/entity/DefaultPersistenceContextTest.java
@@ -3,7 +3,7 @@ package persistence.entity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import persistence.fixture.EntityWithId;
-import persistence.sql.meta.EntityKey;
+import persistence.meta.EntityKey;
 
 import java.lang.reflect.Field;
 import java.util.Map;
