+++ src/test/java/persistence/sql/dml/UpdateQueryBuilderTest.java
@@ -3,7 +3,7 @@ package persistence.sql.dml;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import persistence.fixture.EntityWithId;
-import persistence.sql.meta.EntityTable;
+import persistence.meta.EntityTable;
 
 import static org.assertj.core.api.Assertions.*;
